Update dependencies and add Markdown support
This commit runs a `pipenv update` on the Pipfile to make sure latest
libraries are distributed in the lock file. I also added the
`recommonmark` library to the environment in order to add out-of-the-box
support for Markdown documents in the quickstart. Markdown files are now
dynamically detected and converted into beautiful HTML.

I also set the Pipfile to use Python 3.7 instead of Python 3.6.

Signed-off-by: Justin W. Flory <git@jwf.io>